refactor : FareSectionType 위치 변경
- in domain Package
