+++ src/test/java/nextstep/subway/fare/FareSectionTypeTest.java
@@ -2,6 +2,7 @@ package nextstep.subway.fare;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.subway.fare.domain.FareSectionType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
